update boost version (#1948)
* update boost version to 1.76

* fix compile error caused by boost upgradle

* fix ut compile error caused by boost upgradle

+++ libprecompiled/extension/CryptoPrecompiled.cpp
@@ -95,7 +95,7 @@ PrecompiledExecResult::Ptr CryptoPrecompiled::call(
             break;
         }
         // no defined function
-        callResult->setExecResult(abi.abiIn("", u256(CODE_UNKNOW_FUNCTION_CALL)));
+        callResult->setExecResult(abi.abiIn("", u256((uint32_t)CODE_UNKNOW_FUNCTION_CALL)));
     } while (0);
     return callResult;
 }
